+++ windows/src/desktop/kmshell/main/initprog.pas
@@ -513,7 +513,6 @@ begin
         else ExitCode := 1;
 
     fmInstallTip:
-      // TODO: handle when not enough parameters
       if TTipMaintenance.InstallTip(StrToIntDef('$'+FirstKeyboardFilename, 0), SecondKeyboardFileName, ThirdKeyboardFileName, FourthKeyboardFileName)
         then ExitCode := 0
         else ExitCode := 1;